Factor out extension-init.ts and extension-activation.ts.
