+++ src/client/extension.ts
@@ -1,4 +1,5 @@
 'use strict';
+
 // tslint:disable:no-var-requires no-require-imports
 
 // This line should always be right on top.
@@ -26,95 +27,32 @@ const stopWatch = new StopWatch();
 //===============================================
 // loading starts here
 
-import { Container } from 'inversify';
-import {
-    CodeActionKind,
-    debug,
-    DebugConfigurationProvider,
-    Disposable,
-    ExtensionContext,
-    languages,
-    Memento,
-    OutputChannel,
-    ProgressLocation,
-    ProgressOptions,
-    window
-} from 'vscode';
+import { ProgressLocation, ProgressOptions, window } from 'vscode';
 
-import { registerTypes as activationRegisterTypes } from './activation/serviceRegistry';
-import { IExtensionActivationManager, ILanguageServerExtension } from './activation/types';
 import { buildApi, IExtensionApi } from './api';
-import { registerTypes as appRegisterTypes } from './application/serviceRegistry';
-import { IApplicationDiagnostics } from './application/types';
-import { DebugService } from './common/application/debugService';
-import { IApplicationShell, ICommandManager, IWorkspaceService } from './common/application/types';
-import { Commands, PYTHON, PYTHON_LANGUAGE, STANDARD_OUTPUT_CHANNEL } from './common/constants';
-import { registerTypes as installerRegisterTypes } from './common/installer/serviceRegistry';
+import { IApplicationShell } from './common/application/types';
 import { traceError } from './common/logger';
-import { registerTypes as platformRegisterTypes } from './common/platform/serviceRegistry';
-import { registerTypes as processRegisterTypes } from './common/process/serviceRegistry';
-import { registerTypes as commonRegisterTypes } from './common/serviceRegistry';
-import {
-    GLOBAL_MEMENTO,
-    IAsyncDisposableRegistry,
-    IConfigurationService,
-    IDisposableRegistry,
-    IExperimentsManager,
-    IExtensionContext,
-    IFeatureDeprecationManager,
-    IMemento,
-    IOutputChannel,
-    WORKSPACE_MEMENTO
-} from './common/types';
+import { IAsyncDisposableRegistry, IExtensionContext } from './common/types';
 import { createDeferred } from './common/utils/async';
-import { Common, OutputChannelNames } from './common/utils/localize';
-import { registerTypes as variableRegisterTypes } from './common/variables/serviceRegistry';
-import { JUPYTER_OUTPUT_CHANNEL } from './datascience/constants';
-import { registerTypes as dataScienceRegisterTypes } from './datascience/serviceRegistry';
-import { IDataScience } from './datascience/types';
-import { DebuggerTypeName } from './debugger/constants';
-import { DebugSessionEventDispatcher } from './debugger/extension/hooks/eventHandlerDispatcher';
-import { IDebugSessionEventHandlers } from './debugger/extension/hooks/types';
-import { registerTypes as debugConfigurationRegisterTypes } from './debugger/extension/serviceRegistry';
-import { IDebugConfigurationService, IDebuggerBanner } from './debugger/extension/types';
-import { registerTypes as formattersRegisterTypes } from './formatters/serviceRegistry';
-import { IInterpreterSelector } from './interpreter/configuration/types';
-import {
-    IInterpreterLocatorProgressHandler,
-    IInterpreterLocatorProgressService,
-    IInterpreterService
-} from './interpreter/contracts';
-import { registerTypes as interpretersRegisterTypes } from './interpreter/serviceRegistry';
-import { ServiceContainer } from './ioc/container';
-import { ServiceManager } from './ioc/serviceManager';
-import { IServiceContainer, IServiceManager } from './ioc/types';
-import { getLanguageConfiguration } from './language/languageConfiguration';
-import { LinterCommands } from './linters/linterCommands';
-import { registerTypes as lintersRegisterTypes } from './linters/serviceRegistry';
-import { PythonCodeActionProvider } from './providers/codeActionProvider/pythonCodeActionProvider';
-import { PythonFormattingEditProvider } from './providers/formatProvider';
-import { ReplProvider } from './providers/replProvider';
-import { registerTypes as providersRegisterTypes } from './providers/serviceRegistry';
-import { activateSimplePythonRefactorProvider } from './providers/simpleRefactorProvider';
-import { TerminalProvider } from './providers/terminalProvider';
-import { ISortImportsEditingProvider } from './providers/types';
+import { Common } from './common/utils/localize';
+import { activateComponents } from './extension-activation';
+import { initializeComponents, initializeGlobals } from './extension-init';
+import { IServiceContainer } from './ioc/types';
 import { sendErrorTelemetry, sendStartupTelemetryInBackground } from './startupTelemetry';
-import { registerTypes as commonRegisterTerminalTypes } from './terminals/serviceRegistry';
-import { ICodeExecutionManager, ITerminalAutoActivation } from './terminals/types';
-import { TEST_OUTPUT_CHANNEL } from './testing/common/constants';
-import { ITestContextService } from './testing/common/types';
-import { ITestCodeNavigatorCommandHandler, ITestExplorerCommandHandler } from './testing/navigation/types';
-import { registerTypes as unitTestsRegisterTypes } from './testing/serviceRegistry';
 
 durations.codeLoadingTime = stopWatch.elapsedTime;
 
 //===============================================
 // loading ends here
 
+// These persist between activations:
 const activationDeferred = createDeferred<void>();
-let activatedServiceContainer: ServiceContainer | undefined;
+let activatedServiceContainer: IServiceContainer | undefined;
 
-export async function activate(context: ExtensionContext): Promise<IExtensionApi> {
+/////////////////////////////
+// public functions
+
+export async function activate(context: IExtensionContext): Promise<IExtensionApi> {
     try {
         return await activateUnsafe(context);
     } catch (ex) {
@@ -141,7 +79,7 @@ export function deactivate(): Thenable<void> {
 // activation helpers
 
 // tslint:disable-next-line:max-func-body-length
-async function activateUnsafe(context: ExtensionContext): Promise<IExtensionApi> {
+async function activateUnsafe(context: IExtensionContext): Promise<IExtensionApi> {
     displayProgress(activationDeferred.promise);
     durations.startActivateTime = stopWatch.elapsedTime;
 
@@ -149,6 +87,7 @@ async function activateUnsafe(context: ExtensionContext): Promise<IExtensionApi>
     // activation starts here
 
     const [serviceManager, serviceContainer] = initializeGlobals(context);
+    activatedServiceContainer = serviceContainer;
     initializeComponents(context, serviceManager, serviceContainer);
     const activationPromise = activateComponents(context, serviceManager, serviceContainer);
 
@@ -167,194 +106,6 @@ async function activateUnsafe(context: ExtensionContext): Promise<IExtensionApi>
 function displayProgress(promise: Promise<any>) {
     const progressOptions: ProgressOptions = { location: ProgressLocation.Window, title: Common.loadingExtension() };
     window.withProgress(progressOptions, () => promise);
-}
-
-/////////////////////////////
-// intialization (globals & components)
-
-// The code in this section should do nothing more complex than register
-// objects to DI and simple init (e.g. no side effects).  That implies
-// that constructors are likewise simple and do no work.  It also means
-// that it is inherently synchronous.
-
-function initializeGlobals(context: ExtensionContext): [IServiceManager, IServiceContainer] {
-    const cont = new Container();
-    const serviceManager = new ServiceManager(cont);
-    const serviceContainer = new ServiceContainer(cont);
-    activatedServiceContainer = serviceContainer;
-
-    serviceManager.addSingletonInstance<IServiceContainer>(IServiceContainer, serviceContainer);
-    serviceManager.addSingletonInstance<IServiceManager>(IServiceManager, serviceManager);
-
-    serviceManager.addSingletonInstance<Disposable[]>(IDisposableRegistry, context.subscriptions);
-    serviceManager.addSingletonInstance<Memento>(IMemento, context.globalState, GLOBAL_MEMENTO);
-    serviceManager.addSingletonInstance<Memento>(IMemento, context.workspaceState, WORKSPACE_MEMENTO);
-    serviceManager.addSingletonInstance<IExtensionContext>(IExtensionContext, context);
-
-    return [serviceManager, serviceContainer];
-}
-
-function initializeComponents(
-    _context: ExtensionContext,
-    _serviceManager: IServiceManager,
-    _serviceContainer: IServiceContainer
-) {
-    // We will be pulling code over from activateLegacy().
-}
-
-/////////////////////////////
-// component activation
-
-async function activateComponents(
-    context: ExtensionContext,
-    serviceManager: IServiceManager,
-    serviceContainer: IServiceContainer
-) {
-    // We will be pulling code over from activateLegacy().
-
-    return activateLegacy(context, serviceManager, serviceContainer);
-}
-
-/////////////////////////////
-// old activation code
-
-// tslint:disable-next-line:no-suspicious-comment
-// TODO(GH-10454): Gradually move simple initialization
-// and DI registration currently in this function over
-// to initializeComponents().  Likewise with complex
-// init and activation: move them to activateComponents().
-
-async function activateLegacy(
-    context: ExtensionContext,
-    serviceManager: IServiceManager,
-    serviceContainer: IServiceContainer
-) {
-    registerServices(serviceManager);
-    await initializeServices(context, serviceManager, serviceContainer);
-
-    const manager = serviceContainer.get<IExtensionActivationManager>(IExtensionActivationManager);
-    context.subscriptions.push(manager);
-    const activationPromise = manager.activate();
-
-    serviceManager.get<ITerminalAutoActivation>(ITerminalAutoActivation).register();
-    const configuration = serviceManager.get<IConfigurationService>(IConfigurationService);
-    const pythonSettings = configuration.getSettings();
-
-    const standardOutputChannel = serviceContainer.get<OutputChannel>(IOutputChannel, STANDARD_OUTPUT_CHANNEL);
-    activateSimplePythonRefactorProvider(context, standardOutputChannel, serviceContainer);
-
-    const sortImports = serviceContainer.get<ISortImportsEditingProvider>(ISortImportsEditingProvider);
-    sortImports.registerCommands();
-
-    serviceManager.get<ICodeExecutionManager>(ICodeExecutionManager).registerCommands();
-
-    // At this point we have enough information to send valid telemetry.
-
-    const workspaceService = serviceContainer.get<IWorkspaceService>(IWorkspaceService);
-    const interpreterManager = serviceContainer.get<IInterpreterService>(IInterpreterService);
-    interpreterManager
-        .refresh(workspaceService.hasWorkspaceFolders ? workspaceService.workspaceFolders![0].uri : undefined)
-        .catch(ex => traceError('Python Extension: interpreterManager.refresh', ex));
-
-    // Activate data science features
-    const dataScience = serviceManager.get<IDataScience>(IDataScience);
-    dataScience.activate().ignoreErrors();
-
-    context.subscriptions.push(new LinterCommands(serviceManager));
-
-    languages.setLanguageConfiguration(PYTHON_LANGUAGE, getLanguageConfiguration());
-
-    if (pythonSettings && pythonSettings.formatting && pythonSettings.formatting.provider !== 'internalConsole') {
-        const formatProvider = new PythonFormattingEditProvider(context, serviceContainer);
-        context.subscriptions.push(languages.registerDocumentFormattingEditProvider(PYTHON, formatProvider));
-        context.subscriptions.push(languages.registerDocumentRangeFormattingEditProvider(PYTHON, formatProvider));
-    }
-
-    const deprecationMgr = serviceContainer.get<IFeatureDeprecationManager>(IFeatureDeprecationManager);
-    deprecationMgr.initialize();
-    context.subscriptions.push(deprecationMgr);
-
-    context.subscriptions.push(new ReplProvider(serviceContainer));
-
-    const terminalProvider = new TerminalProvider(serviceContainer);
-    terminalProvider.initialize(window.activeTerminal).ignoreErrors();
-    context.subscriptions.push(terminalProvider);
-
-    context.subscriptions.push(
-        languages.registerCodeActionsProvider(PYTHON, new PythonCodeActionProvider(), {
-            providedCodeActionKinds: [CodeActionKind.SourceOrganizeImports]
-        })
-    );
-
-    serviceContainer.getAll<DebugConfigurationProvider>(IDebugConfigurationService).forEach(debugConfigProvider => {
-        context.subscriptions.push(debug.registerDebugConfigurationProvider(DebuggerTypeName, debugConfigProvider));
-    });
-
-    serviceContainer.get<IDebuggerBanner>(IDebuggerBanner).initialize();
-
-    return activationPromise;
-}
-
-function registerServices(serviceManager: IServiceManager) {
-    const standardOutputChannel = window.createOutputChannel(OutputChannelNames.python());
-    const unitTestOutChannel = window.createOutputChannel(OutputChannelNames.pythonTest());
-    const jupyterOutputChannel = window.createOutputChannel(OutputChannelNames.jupyter());
-    serviceManager.addSingletonInstance<OutputChannel>(IOutputChannel, standardOutputChannel, STANDARD_OUTPUT_CHANNEL);
-    serviceManager.addSingletonInstance<OutputChannel>(IOutputChannel, unitTestOutChannel, TEST_OUTPUT_CHANNEL);
-    serviceManager.addSingletonInstance<OutputChannel>(IOutputChannel, jupyterOutputChannel, JUPYTER_OUTPUT_CHANNEL);
-
-    commonRegisterTypes(serviceManager);
-    processRegisterTypes(serviceManager);
-    variableRegisterTypes(serviceManager);
-    unitTestsRegisterTypes(serviceManager);
-    lintersRegisterTypes(serviceManager);
-    interpretersRegisterTypes(serviceManager);
-    formattersRegisterTypes(serviceManager);
-    platformRegisterTypes(serviceManager);
-    installerRegisterTypes(serviceManager);
-    commonRegisterTerminalTypes(serviceManager);
-    dataScienceRegisterTypes(serviceManager);
-    debugConfigurationRegisterTypes(serviceManager);
-
-    const configuration = serviceManager.get<IConfigurationService>(IConfigurationService);
-    const languageServerType = configuration.getSettings().languageServer;
-
-    appRegisterTypes(serviceManager, languageServerType);
-    providersRegisterTypes(serviceManager);
-    activationRegisterTypes(serviceManager, languageServerType);
-}
-
-async function initializeServices(
-    context: ExtensionContext,
-    serviceManager: IServiceManager,
-    serviceContainer: IServiceContainer
-) {
-    const abExperiments = serviceContainer.get<IExperimentsManager>(IExperimentsManager);
-    await abExperiments.activate();
-    const selector = serviceContainer.get<IInterpreterSelector>(IInterpreterSelector);
-    selector.initialize();
-    context.subscriptions.push(selector);
-
-    const interpreterManager = serviceContainer.get<IInterpreterService>(IInterpreterService);
-    interpreterManager.initialize();
-
-    const handlers = serviceManager.getAll<IDebugSessionEventHandlers>(IDebugSessionEventHandlers);
-    const disposables = serviceManager.get<IDisposableRegistry>(IDisposableRegistry);
-    const dispatcher = new DebugSessionEventDispatcher(handlers, DebugService.instance, disposables);
-    dispatcher.registerEventHandlers();
-
-    const cmdManager = serviceContainer.get<ICommandManager>(ICommandManager);
-    const outputChannel = serviceManager.get<OutputChannel>(IOutputChannel, STANDARD_OUTPUT_CHANNEL);
-    disposables.push(cmdManager.registerCommand(Commands.ViewOutput, () => outputChannel.show()));
-
-    // Display progress of interpreter refreshes only after extension has activated.
-    serviceContainer.get<IInterpreterLocatorProgressHandler>(IInterpreterLocatorProgressHandler).register();
-    serviceContainer.get<IInterpreterLocatorProgressService>(IInterpreterLocatorProgressService).register();
-    serviceContainer.get<IApplicationDiagnostics>(IApplicationDiagnostics).register();
-    serviceContainer.get<ITestCodeNavigatorCommandHandler>(ITestCodeNavigatorCommandHandler).register();
-    serviceContainer.get<ITestExplorerCommandHandler>(ITestExplorerCommandHandler).register();
-    serviceContainer.get<ILanguageServerExtension>(ILanguageServerExtension).register();
-    serviceContainer.get<ITestContextService>(ITestContextService).register();
 }
 
 /////////////////////////////
